feat: Remove clouddrive-connectors portlet from bottom-all-container configuration - EXO-65110 (#25)
Prior to this change, in bottom-all-container, we inject and load clouddrive-connectors extension as a portlet, so, it is loaded on all pages even if no need to be used which can lead to loading performance issues.
To fix that, we must use js load-group, which are now available through all webapps to load clouddrive-connectors extension only with the portlet CloudDriveSettingsPortlet.